+++ Libft/ft_puttab.c
@@ -6,7 +6,7 @@
 /*   By: sklepper <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2018/04/07 16:42:37 by sklepper          #+#    #+#             */
-/*   Updated: 2018/04/16 15:50:59 by sklepper         ###   ########.fr       */
+/*   Updated: 2018/04/22 01:48:32 by nihuynh          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,13 +14,6 @@
 
 void	ft_puttab(char **tab)
 {
-	int i;
-
-	i = 0;
-	while (tab[i] != 0)
-	{
-		ft_putstr(tab[i]);
-		ft_putchar('\n');
-		i++;
-	}
+	while (*tab)
+		ft_putendl(*tab++);
 }