+++ cmake-build-debug/CMakeFiles/Libft.dir/cmake_clean.cmake
@@ -1,5 +1,0 @@
-
-# Per-language clean rules from dependency scanning.
-foreach(lang )
-  include(CMakeFiles/Libft.dir/cmake_clean_${lang}.cmake OPTIONAL)
-endforeach()